LPS-130175 Move GlobalContext component
