+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/App.scss
@@ -6,7 +6,6 @@
 @import 'DragPreview';
 @import 'DropZone';
 @import 'EditModeSelector';
-@import 'GlobalContext';
 @import 'Layout';
 @import 'LayoutViewport';
 @import 'NetworkstatusBar';
@@ -23,6 +22,8 @@
 @import 'layout-data-items/RowWithControls';
 
 @import 'undo/UndoHistory';
+
+@import '../contexts/GlobalContext';
 
 @import '../../common/components/Button';
 @import '../../common/components/Collapse';
