+++ modules/apps/layout/layout-content-page-editor-web/test/page_editor/app/components/ImageSelectorSize.test.js
@@ -16,13 +16,13 @@ import '@testing-library/jest-dom/extend-expect';
 import {act, cleanup, render} from '@testing-library/react';
 import React from 'react';
 
-import {useGlobalContext} from '../../../../src/main/resources/META-INF/resources/page_editor/app/components/GlobalContext';
+import {useGlobalContext} from '../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/GlobalContext';
 import ImageService from '../../../../src/main/resources/META-INF/resources/page_editor/app/services/ImageService';
 import {StoreAPIContextProvider} from '../../../../src/main/resources/META-INF/resources/page_editor/app/store';
 import {ImageSelectorSize} from '../../../../src/main/resources/META-INF/resources/page_editor/common/components/ImageSelectorSize';
 
 jest.mock(
-	'../../../../src/main/resources/META-INF/resources/page_editor/app/components/GlobalContext'
+	'../../../../src/main/resources/META-INF/resources/page_editor/app/contexts/GlobalContext'
 );
 
 jest.mock(
